+++ Proiect mds/Proiect mds/Controllers/UserController.cs
@@ -400,80 +400,7 @@ namespace Proiect_mds.Controllers
         }
 
 
-        public ActionResult AccesezSpanzurat(string email)
-        {
-            using (DBUsersEntities1 db = new DBUsersEntities1())
-            {
-                db.Jocuris.FirstOrDefault(a => a.Nume == "Spanzurat").Nr_accesari++;
-                db.SaveChanges();
-                return RedirectToAction("Game", "Home");
-            }
-        }
-
-        public ActionResult AccesezUrsi(string email)
-        {
-            using (DBUsersEntities1 db = new DBUsersEntities1())
-            {
-                db.Jocuris.FirstOrDefault(a => a.Nume == "Ursi").Nr_accesari++;
-                db.SaveChanges();
-                return RedirectToAction("Bears", "Home");
-            }
-        }
-
-        public ActionResult AccesezInvadez(string email)
-        {
-            using (DBUsersEntities1 db = new DBUsersEntities1())
-            {
-                db.Jocuris.FirstOrDefault(a => a.Nume == "Invadez").Nr_accesari++;
-                db.SaveChanges();
-                return RedirectToAction("SpaceInvaders", "Home");
-            }
-        }
-
-        public ActionResult AccesezFlap(string email)
-        {
-            using (DBUsersEntities1 db = new DBUsersEntities1())
-            {
-                db.Jocuris.FirstOrDefault(a => a.Nume == "Flap").Nr_accesari++;
-                db.SaveChanges();
-                return RedirectToAction("Flappy", "Home");
-            }
-        }
-
-        public ActionResult AccesezMemo(string email)
-        {
-            using (DBUsersEntities1 db = new DBUsersEntities1())
-            {
-                db.Jocuris.FirstOrDefault(a => a.Nume == "Memo").Nr_accesari++;
-                db.SaveChanges();
-                return RedirectToAction("MemoryGame", "Home");
-            }
-        }
-
-        public ActionResult AccesezSnek(string email)
-        {
-            using (DBUsersEntities1 db = new DBUsersEntities1())
-            {
-                db.Jocuris.FirstOrDefault(a => a.Nume == "Snek").Nr_accesari++;
-                db.SaveChanges();
-                return RedirectToAction("Snake", "Home");
-            }
-        }
-
-        static public string Top3()
-        {
-            using (DBUsersEntities1 db = new DBUsersEntities1())
-            {
-                string result = "";
-                var v = db.Jocuris.OrderByDescending(a => a.Nr_accesari);//ForEach(a => System.Diagnostics.Debug.Print(a.Nume));
-                System.Collections.Generic.List<Jocuri> l = v.ToList();
-                l.ForEach(a => result = result + " " + a.Nume);
-                string[] vector = result.Split(' ');
-                return (vector[1] + " " + vector[2] + " " + vector[3]);
-            }
-        }
-
-
+
