+++ Proiect mds/Proiect mds/Views/Home/Index.cshtml
@@ -30,7 +30,7 @@
             }    
      }
     </a>
-    <a href='@Url.Action( "AccesezSpanzurat", "User")'>
+    <a href='@Url.Action( "AccesezSpanzurat", "Joc")'>
         <img src="~/Content/Images/SPANZURAT.png" height="350" width="350" />
     </a>
 
@@ -48,7 +48,7 @@
             }
         }
     </a>
-    <a href='@Url.Action( "AccesezUrsi", "User")'>
+    <a href='@Url.Action( "AccesezUrsi", "Joc")'>
         <img src="~/Content/Images/URS.png" height="350" width="350">
     </a>
 
@@ -68,7 +68,7 @@
         }
     </a>
 
-     <a href='@Url.Action( "AccesezMemo", "User")'>
+     <a href='@Url.Action( "AccesezMemo", "Joc")'>
           <img src="~/Content/Images/MEMO.png" height="350" width="350">
      </a>
         <br>
@@ -89,7 +89,7 @@
     }
 
     </a>
-        <a href='@Url.Action( "AccesezFlap", "User")'>
+        <a href='@Url.Action( "AccesezFlap", "Joc")'>
             <img src="~/Content/Images/FLAP.png" height="350" width="350">
         </a>
 
@@ -108,7 +108,7 @@
         }
 
      </a>
-        <a href='@Url.Action( "AccesezInvadez", "User")'>
+        <a href='@Url.Action( "AccesezInvadez", "Joc")'>
             <img src="~/Content/Images/INVADEZ.png" height="350" width="350">
         </a>
 
@@ -125,7 +125,7 @@
             }
         }
     </a>
-        <a href='@Url.Action( "AccesezSnek", "User")'>
+        <a href='@Url.Action( "AccesezSnek", "Joc")'>
             <img src="~/Content/Images/SNEK.png" height="350" width="350">
         </a>
 </div>
@@ -142,15 +142,3 @@
 
 }    
 
-
-
-<script>
-
-    function place(img)
-    {
-                                                     
-        img.src = '/Content/Images/like.png';    
-    }
-         
-
-</script>
